requirements.txt
# --- core runtime ---
numpy==1.26.4
pillow==11.3.0
PyYAML==6.0.2
requests==2.32.5
tqdm==4.67.1
filelock==3.19.1
fsspec==2025.9.0
regex==2025.9.1
safetensors==0.6.2

# --- torch stack ---
torch==2.8.0
torchvision==0.23.0

# --- huggingface ecosystem ---
tokenizers==0.22.0
huggingface-hub==0.34.4
transformers==4.56.1

# --- classification/segmentation backbones ---
timm==1.0.19
segmentation_models_pytorch==0.5.0

# --- cv/io ---
opencv-python-headless==4.11.0.86
imageio==2.37.0
tifffile==2025.8.28

# --- science stack (used by skimage/sklearn) ---
scipy==1.16.1
scikit-image==0.25.2
scikit-learn==1.7.1
threadpoolctl==3.6.0

# --- coco/detectron2 deps (must come before detectron2) ---
pycocotools==2.0.10
yacs==0.1.8
iopath==0.1.10
fvcore==0.1.5.post20221221
tabulate==0.9.0
termcolor==3.1.0

# --- detectron2 (git) ---
#detectron2 @ git+https://github.com/facebookresearch/detectron2.git@a1ce2f956a1d2212ad672e3c47d53405c2fe4312

# --- SAM / SAM-2 (git; require torch/torchvision above) ---
#SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
#segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf

# --- config (optional but present in your env) ---
hydra-core==1.3.2
omegaconf==2.3.0

# --- logging/visualization (optional) ---
tensorboard==2.20.0
